new year rebuild  - rewrite the extract_vars() function to get environment variables for the cert-service  - fix bug in the icinga2 smashing job

+++ rootfs/init/icinga_cert.sh
@@ -1,6 +1,13 @@
 
 ICINGA_API_PORT=${ICINGA_API_PORT:-5665}
-ICINGA_CERT_SERVICE=${ICINGA_CERT_SERVICE:-false}
+USE_CERT_SERVICE=${USE_CERT_SERVICE:-'false'}
+
+ICINGA_CERT_SERVICE_SERVER=
+ICINGA_CERT_SERVICE_PORT=
+ICINGA_CERT_SERVICE_BA_USER=
+ICINGA_CERT_SERVICE_BA_PASSWORD=
+ICINGA_CERT_SERVICE_API_USER=
+ICINGA_CERT_SERVICE_API_PASSWORD=
 
 # get a new icinga certificate from our icinga-master
 #
@@ -9,12 +16,12 @@ get_certificate() {
 
   validate_local_ca
 
-  if [ -f ${WORK_DIR}/pki/${HOSTNAME}.key ]
+  if [ -f ${WORK_DIR}/pki/${HOSTNAME}/${HOSTNAME}.key ]
   then
     return
   fi
 
-  if [ ${ICINGA_CERT_SERVICE} ]
+  if [ "${USE_CERT_SERVICE}" == "true" ]
   then
     echo ""
     echo " [i] we ask our cert-service for a certificate .."
@@ -26,7 +33,7 @@ get_certificate() {
       --silent \
       --request GET \
       --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-      --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+      --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
       --write-out "%{http_code}\n" \
       --output /tmp/request_${HOSTNAME}.json \
       http://${ICINGA_CERT_SERVICE_SERVER}:${ICINGA_CERT_SERVICE_PORT}${ICINGA_CERT_SERVICE_PATH}v2/request/${HOSTNAME})
@@ -44,6 +51,8 @@ get_certificate() {
 
       mkdir -p ${WORK_DIR}/pki/${HOSTNAME}
 
+      cp -a /tmp/request_${HOSTNAME}.json ${WORK_DIR}/pki/${HOSTNAME}/
+
       # get our created cert
       #
       code=$(curl \
@@ -51,7 +60,7 @@ get_certificate() {
         --silent \
         --request GET \
         --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-        --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+        --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
         --header "X-CHECKSUM: ${checksum}" \
         --write-out "%{http_code}\n" \
         --request GET \
@@ -86,7 +95,7 @@ get_certificate() {
       else
         echo " [E] can't download out certificate!"
 
-        rm -rf ${WORK_DIR}/pki 2> /dev/null
+        rm -rf ${WORK_DIR}/pki/${HOSTNAME} 2> /dev/null
 
         unset ICINGA_API_PKI_PATH
 
@@ -94,12 +103,17 @@ get_certificate() {
       fi
     else
 
-      error=$(cat /tmp/request_${HOSTNAME}.json)
+      if [ -f /tmp/request_${HOSTNAME}.json ]
+      then
+        error=$(cat /tmp/request_${HOSTNAME}.json)
 
-      echo " [E] ${code} - the cert-service tell us a problem: '${error}'"
-      echo " [E] exit ..."
+        echo " [E] ${code} - the cert-service tell us a problem: '${error}'"
+        echo " [E] exit ..."
 
-      rm -f /tmp/request_${HOSTNAME}.json
+        rm -f /tmp/request_${HOSTNAME}.json
+      else
+        echo " [E] ${code} - the cert-service has an unknown problem."
+      fi
       exit 1
     fi
 
@@ -125,7 +139,7 @@ validate_local_ca() {
       --silent \
       --request GET \
       --header "X-API-USER: ${ICINGA_CERT_SERVICE_API_USER}" \
-      --header "X-API-KEY: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
+      --header "X-API-PASSWORD: ${ICINGA_CERT_SERVICE_API_PASSWORD}" \
       --write-out "%{http_code}\n" \
       --output /tmp/validate_ca_${HOSTNAME}.json \
       http://${ICINGA_CERT_SERVICE_SERVER}:${ICINGA_CERT_SERVICE_PORT}${ICINGA_CERT_SERVICE_PATH}v2/validate/${checksum})
@@ -149,6 +163,85 @@ validate_local_ca() {
   fi
 }
 
+
+extract_vars() {
+
+  if [ ! -z "${ICINGA_CERT_SERVICE}" ]
+  then
+
+    USE_FALLBACK="false"
+
+    if ( [ "${ICINGA_CERT_SERVICE}" != "true" ] && [ "${ICINGA_CERT_SERVICE}" != "false" ] )
+    then
+      echo "${ICINGA_CERT_SERVICE}" | json_verify -q 2> /dev/null
+
+      if [ $? -gt 0 ]
+      then
+        echo " [W] the ICINGA_CERT_SERVICE Environment is not an json"
+        echo " [W] use fallback strategy."
+        USE_FALLBACK="true"
+      fi
+    else
+      echo " [W] the ICINGA_CERT_SERVICE Environment is not an json"
+      echo " [W] use fallback strategy."
+      USE_FALLBACK="true"
+    fi
+
+
+    if [ "${USE_FALLBACK}" == "false" ]
+    then
+
+      if ( [ "${ICINGA_CERT_SERVICE}" == "true" ] || [ "${ICINGA_CERT_SERVICE}" == "false" ] )
+      then
+        echo " [E] the ICINGA_CERT_SERVICE Environment must be an json, not true or false!"
+        exit 1
+      fi
+
+      ICINGA_CERT_SERVICE_SERVER=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .server)
+      ICINGA_CERT_SERVICE_PORT=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .port)
+      ICINGA_CERT_SERVICE_PATH=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .path)
+      ICINGA_CERT_SERVICE_API_USER=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .api.user)
+      ICINGA_CERT_SERVICE_API_PASSWORD=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .api.password)
+      ICINGA_CERT_SERVICE_BA_USER=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .ba.user)
+      ICINGA_CERT_SERVICE_BA_PASSWORD=$(echo "${ICINGA_CERT_SERVICE}" | jq --raw-output .ba.password)
+
+      [ "${ICINGA_CERT_SERVICE_SERVER}" == null ] && ICINGA_CERT_SERVICE_SERVER=
+      [ "${ICINGA_CERT_SERVICE_PORT}" == null ] && ICINGA_CERT_SERVICE_PORT=8080
+      [ "${ICINGA_CERT_SERVICE_PATH}" == null ] && ICINGA_CERT_SERVICE_PATH="/"
+      [ "${ICINGA_CERT_SERVICE_API_USER}" == null ] && ICINGA_CERT_SERVICE_API_USER=
+      [ "${ICINGA_CERT_SERVICE_API_PASSWORD}" == null ] && ICINGA_CERT_SERVICE_API_PASSWORD=
+      [ "${ICINGA_CERT_SERVICE_BA_USER}" == null ] && ICINGA_CERT_SERVICE_BA_USER=
+      [ "${ICINGA_CERT_SERVICE_BA_PASSWORD}" == null ] && ICINGA_CERT_SERVICE_BA_PASSWORD=
+
+    else
+
+      ICINGA_CERT_SERVICE_SERVER=${ICINGA_CERT_SERVICE_SERVER:-}
+      ICINGA_CERT_SERVICE_PORT=${ICINGA_CERT_SERVICE_PORT:-8080}
+      ICINGA_CERT_SERVICE_PATH=${ICINGA_CERT_SERVICE_PATH:-'/'}
+      ICINGA_CERT_SERVICE_API_USER=${ICINGA_CERT_SERVICE_API_USER:-''}
+      ICINGA_CERT_SERVICE_API_PASSWORD=${ICINGA_CERT_SERVICE_API_PASSWORD:-''}
+      ICINGA_CERT_SERVICE_BA_USER=${ICINGA_CERT_SERVICE_BA_USER:-"admin"}
+      ICINGA_CERT_SERVICE_BA_PASSWORD=${ICINGA_CERT_SERVICE_BA_PASSWORD:-"admin"}
+    fi
+  fi
+
+
+  if (
+    [ ! -z ${ICINGA_CERT_SERVICE_SERVER} ] &&
+    [ ! -z ${ICINGA_CERT_SERVICE_PORT} ] &&
+    [ ! -z ${ICINGA_CERT_SERVICE_BA_USER} ] &&
+    [ ! -z ${ICINGA_CERT_SERVICE_BA_PASSWORD} ] &&
+    [ ! -z ${ICINGA_CERT_SERVICE_API_USER} ] &&
+    [ ! -z ${ICINGA_CERT_SERVICE_API_PASSWORD} ]
+  )
+  then
+    USE_CERT_SERVICE="true"
+  else
+    USE_CERT_SERVICE="false"
+  fi
+}
+
+extract_vars
 . /init/wait_for/cert_service.sh
 get_certificate
